+++ test/envoy/core_test.clj
@@ -19,9 +19,21 @@
 (defenv :envoy-bad-type
   "Re-declaring a variable.")
 
+(defenv :envoy-custom-parser
+  "Specifying a special parsing fn."
+  :parser clojure.edn/read-string)
+
 
 (deftest env-declaration
   (is (contains? env/known-vars :envoy-test))
   (is (= "Test environment variable." (get-in env/known-vars [:envoy-test :description])))
   (is (= :integer (get-in env/known-vars [:envoy-test :type])))
   (is (= 'envoy.core-test (get-in env/known-vars [:envoy-test :ns]))))
+
+(deftest parser-declaration
+  (do
+    (env/set-env! :envoy-custom-parser "{:a 1}")
+    (is (= {:a 1} (env/env :envoy-custom-parser))))
+  (do
+    (env/set-env! :envoy-custom-parser "{:a {:b [\"1\" 2.0 \\3 \" \"]}}")
+    (is (= {:a {:b ["1" 2.0 \3 " "]}} (env/env :envoy-custom-parser)))))