+++ CHANGELOG.md
@@ -7,7 +7,15 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-...
+## [0.3.2] - 2018-09-07
+
+### Added
+- Add a `:parser` field to `defenv` that sets a custom function as the
+  value parser.
+
+### Fixed
+- Fixed bug preventing `EnvironmentMap` from being used with `find`
+  and `select-keys`.
 
 ## [0.3.1] - 2017-02-02
 
@@ -46,7 +54,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Initial release.
 
-[Unreleased]: https://github.com/amperity/envoy/compare/0.3.1...HEAD
+[Unreleased]: https://github.com/amperity/envoy/compare/0.3.2...HEAD
+[0.3.2]: https://github.com/amperity/envoy/compare/0.3.1...0.3.2
 [0.3.1]: https://github.com/amperity/envoy/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/amperity/envoy/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/amperity/envoy/compare/0.2.0...0.2.1